Changed to log and added time
Added measure of time to calculate pi values and changed the scale of the number of samples to calculate pi to logarithmic, so it does a lot of calculations with few samples and less with larger samples. This reduces the calculation time by a lot while maintaining the graph more or less the same, since there are so many calculations.