Made faster calculation algorithm
I changed the algorithm so that it does not start calculating from scratch each time, but instead uses the previous calculations and build on them. It is, indeed, much faster, so much so that it can reach 10^7 samples in 34s. However, the calculation is also much more imprecise. Because it reuses previous data, the new data is very similar to it, and so it takes a huge number of samples to converge to a more accurate approximation of pi.