Far more accurate value for pi by calculating mean
It the mean of all the approximations of pi to give a far more accurate value. It also gives more information, like the average of all the approximations that are within one and two standard deviations (although it looks like calculating the mean directly is usually better). Also, now it shows the time that the entire code has taken to run, not how long each calculation takes, as it is always the same and makes the console messy.