Create monte_carlo
This file does not generate any graphs, so you can see the code is much simpler. I have also added a progress bar. The aim of this file is that each time it runs, it adds its new result to a file, so that we can get more and more accurate results without having to calculate them at once.